// patientor/backend/src/types.ts
export interface Diagnosis {
  code: string;
  name: string;
  latin?: string
}

export enum HealthCheckRating {
  "Healthy" = 0,
  "LowRisk" = 1,
  "HighRisk" = 2,
  "CriticalRisk" = 3,
}

export interface BaseEntry {
  id: string;
  description: string;
  date: string;
  specialist: string;
  diagnosisCodes?: Array<Diagnosis['code']>;
}

export interface HealthCheckEntry extends BaseEntry {
  type: "HealthCheck";
  healthCheckRating: HealthCheckRating;
}
export interface SickLeave {
  startDate: string;
  endDate: string;
}
export interface OccupationalHealthcareEntry extends BaseEntry {
  type: "OccupationalHealthcare";
  employerName: string;
  sickLeave?: SickLeave;
}

export interface Discharge {
  date: string;
  criteria: string;


}
export interface Hospital extends BaseEntry {
  type: "Hospital";
  discharge: Discharge
}
export type Entry = HealthCheckEntry | OccupationalHealthcareEntry | Hospital;

export enum Gender {
  Male = 'male',
  Female = 'female',
  Other = 'other'
}
export interface Patient {
  id: string;
  name: string;
  dateOfBirth: string;
  ssn: string;
  gender: Gender;
  occupation: string;
  entries: Entry[]

}

type UnionOmit<T, K extends string | number | symbol> = T extends unknown ? Omit<T, K> : never;

export type EntryWithoutId = UnionOmit<Entry, 'id'>;



export type newPatientNoID = Omit<Patient, 'id'>;
export type noSSN = Omit<Patient, 'ssn' | 'entries'>;